+++ HISTORY.md
@@ -56,3 +56,9 @@
 ## 0.3.11
 
 - get symbol from chainId
+- AddressLink component added
+
+## 0.4.0
+
+- New login component added (LoginV2)
+- Network name will be visible in the old login component